requirements.txt
openai>=1.0.0
plotly>=5.20
chainlit==1.1.402
python-dotenv>=1.0
pydantic==2.6.4